requirements.txt
omegaconf==2.0.6
numpy==1.19.2
hydra_core==1.0.6
scikit_learn==0.24.1
